Modularize the debian and arch as a plugin of cupper
