+++ lib/cupper.rb
@@ -6,9 +6,15 @@ require 'cupper/collect'
 require 'cupper/cookbook'
 require 'cupper/entity'
 
+
 module Cupper
   CUPPER_ROOT       = File.expand_path(File.dirname(__FILE__))
   OHAI_PLUGINS_PATH = CUPPER_ROOT + '/cupper/plugins/ohai'
+  CUPPER_PLUGINS_PATH = CUPPER_ROOT + '/cupper/plugins/cupper'
   TEMPLATE_PATH     = CUPPER_ROOT + '/cupper/templates'
+
+  # REVIEW: maybe there is a better way and place to load all plugins
+  # Loading all cupper plugins
+  Dir["#{CUPPER_PLUGINS_PATH}/*.rb"].each { |file| require file }
 end
