+++ lib/cupper/collect.rb
@@ -3,7 +3,6 @@ require 'cupper/ohai_plugins'
 
 module Cupper
   class Collect
-
     ADDITIONAL_OHAI_PLUGINS = [
       'packages',
       'platform_family',
@@ -47,129 +46,6 @@ module Cupper
 
     def data
       @data_extraction
-    end
-  end
-
-  # REVIEW: This should be review in some point in the future
-  # The PlatformCollector is a module that defines all the methods used to extract
-  #   the information about the platform. It is a little pointless in terms of implementation
-  #   but is good to know what method should be implemented.
-  module PlatformCollector
-
-    def packages
-      raise NotImplementedError
-    end
-
-    def links
-      raise NotImplementedError
-    end
-    
-    def services
-      raise NotImplementedError
-    end
-
-    def users
-      raise NotImplementedError
-    end
-
-    def groups
-      raise NotImplementedError
-    end
-
-    def executes
-      raise NotImplementedError
-    end
-
-    def directory
-      raise NotImplementedError
-    end
-
-    def files
-      raise NotImplementedError
-    end
-
-    def templates
-      raise NotImplementedError
-    end
-  end
-
-  class Debian
-    include PlatformCollector
-    def packages(data_extraction)
-      packages = Array.new
-      data_extraction['packages']['packages'].each do |pkg|
-        packages.push(pkg)
-      end
-      packages
-    end
-
-    def files(data_extraction)
-      files = Array.new
-      data_extraction['files']['files'].each do |file|
-        files.push(file)
-      end
-      files
-    end
-
-    def services(data_extraction)
-      services = Array.new
-      data_extraction['services']['services'].each do |service|
-        services.push(service)
-      end
-    end
-
-    def users(data_extraction)
-      users = Array.new
-      data_extraction['etc']['etc']['passwd'].each do |user|
-        users.push(user)
-      end
-    end
-
-    def groups(data_extraction)
-      groups = Array.new
-      data_extraction['etc']['etc']['group'].each do |group|
-        groups.push(group)
-      end
-    end
-  end
-
-  class Arch
-    include PlatformCollector
-    def packages(data_extraction)
-      packages = Array.new
-      data_extraction['pacman']['pacman'].each do |pkg|
-        packages.push(pkg)
-      end
-      packages
-    end
-
-    def files(data_extraction)
-      files = Array.new
-      data_extraction['files']['files'].each do |file|
-        files.push(file)
-      end
-      files
-    end
-
-    def services(data_extraction)
-      services = Array.new
-      data_extraction['services']['services'].each do |service|
-        services.push(service)
-      end
-    end
-
-    def users(data_extraction)
-      users = Array.new
-      data_extraction['etc']['etc']['passwd'].each do |user|
-        users.push(user)
-      end
-    end
-
-    def groups(data_extraction)
-      groups = Array.new
-      data_extraction['etc']['etc']['group'].each do |group|
-        groups.push(group)
-      end
     end
   end
 end
